create ci iam user and deployment role (#63)
Create CI user which can sts assume a deployment role. This role is limited to creating roles, editing the dynamo state lock files and infra within the existing s3 data lake and remote state bucket. The role **does not ** create the initial state infra, this is to limit the scope of privileges.